loadtest/requirements.txt
locust>=2.0.0
grpcio>=1.30.0
grpcio-tools>=1.30.0

